requirements.txt
feedparser==6.0.11
google-generativeai==0.8.3
gspread==6.1.2
google-auth==2.35.0
google-auth-oauthlib==1.2.1
openpyxl==3.1.5
requests==2.32.3
python-dotenv==1.0.1
